+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.9.6
+##### Enhancement
+* Add Error Handler to serve(urlString:...)  
+  [Toshihiro Suzuki](https://github.com/toshi0383)
+  [#113](https://github.com/toshi0383/TVMLKitchen/pull/113)
+
 ## 0.9.5
 ##### Bugfix
 * [Multi Window] Fixed crashing bug when emptying Kitchen.navigationController during Kitchen.window's presentation